New upstream version 3.0.0

+++ src/tss2-fapi/ifapi_io.c
@@ -255,7 +255,7 @@ ifapi_io_check_file_writeable(
     const char *file)
 {
     /* Check access rights to file  */
-    if (access(file, W_OK)) {
+    if (access(file, FAPI_WRITE)) {
         return_error2(TSS2_FAPI_RC_IO_ERROR, "File %s is not writeable.", file);
     }
     return TSS2_RC_SUCCESS;
@@ -272,7 +272,7 @@ ifapi_io_check_file_writeable(
  */
 TSS2_RC
 ifapi_io_check_create_dir(
-    const char *dirname)
+    const char *dirname, int mode)
 {
     TSS2_RC r;
     struct stat fbuffer;
@@ -288,8 +288,9 @@ ifapi_io_check_create_dir(
     }
 
     /* Check access rights to dirname */
-    if (access(dirname, W_OK)) {
-        return_error2(TSS2_FAPI_RC_IO_ERROR, "Directory %s is not writeable.", dirname);
+    if (access(dirname, mode)) {
+        return_error2(TSS2_FAPI_RC_IO_ERROR, "Directory %s is not %s.", dirname,
+                      (mode == FAPI_WRITE) ? "writeable" : "readable");
     }
 
     return TSS2_RC_SUCCESS;
@@ -314,18 +315,26 @@ ifapi_io_remove_file(const char *file)
 /** Remove a directory recursively; i.e. including its subdirectories.
  *
  * @param[in] dirname The directory to be removed
+ * @param[in] keystore_path The path of the current keystore directory, which should
+ *            not be removed.
+ * @param[in] sub_dir The path of a sub directory of the keystore directory,
+ *            which should not be removed (can be NULL, if not needed).
+ *            It must start with a slash.
  * @retval TSS2_RC_SUCCESS if the directories were successfully removed
  * @retval TSS2_FAPI_RC_IO_ERROR if an I/O error occurred
  * @retval TSS2_FAPI_RC_MEMORY: if memory could not be allocated to hold the read data.
  */
 TSS2_RC
 ifapi_io_remove_directories(
-    const char *dirname)
+    const char *dirname,
+    const char *keystore_path,
+    const char *sub_dir)
 {
     DIR *dir;
     struct dirent *entry;
     TSS2_RC r;
     char *path;
+    size_t len_kstore_path, len_dir_path, diff_len, pos;
 
     LOG_TRACE("Removing directory: %s", dirname);
 
@@ -347,7 +356,7 @@ ifapi_io_remove_directories(
             r = ifapi_asprintf(&path, "%s/%s", dirname, entry->d_name);
             goto_if_error(r, "Out of memory", error_cleanup);
 
-            r = ifapi_io_remove_directories(path);
+            r = ifapi_io_remove_directories(path, keystore_path, sub_dir);
             free(path);
             goto_if_error(r, "remove directories.", error_cleanup);
 
@@ -358,7 +367,7 @@ ifapi_io_remove_directories(
         r = ifapi_asprintf(&path, "%s/%s", dirname, entry->d_name);
         goto_if_error(r, "Out of memory", error_cleanup);
 
-        LOG_WARNING("Found a file in directory; removing: %s", path);
+        LOG_WARNING("Removing: %s", path);
 
         if (remove(path) != 0) {
             free(path);
@@ -369,10 +378,22 @@ ifapi_io_remove_directories(
         free(path);
     }
     closedir(dir);
-    LOG_TRACE("Removing target directory %s", dirname);
 
-    if (rmdir(dirname) != 0)
-        return_error2(TSS2_FAPI_RC_IO_ERROR, "Removing directory: %s", dirname);
+    /* Check whether current directory is a keystore directory. These directories should
+       not be deleted. */
+    len_kstore_path = strlen(keystore_path);
+    len_dir_path = strlen(dirname);
+    diff_len = len_dir_path - len_kstore_path;
+    if (diff_len > 1) {
+        pos = strlen(keystore_path);
+        if (keystore_path[pos - 1] == '/')
+            pos += 1;
+        /* Check whether current sub_dir of keystore path should not be deleted. */
+        if (!sub_dir || strcmp(&dirname[pos], sub_dir) != 0) {
+            if (rmdir(dirname) != 0)
+                return_error2(TSS2_FAPI_RC_IO_ERROR, "Removing directory: %s", dirname);
+        }
+    }
 
     LOG_TRACE("SUCCESS");
     return TSS2_RC_SUCCESS;
@@ -408,7 +429,7 @@ ifapi_io_dirfiles(
     check_not_null(files);
     check_not_null(numfiles);
 
-    LOG_TRACE("Removing directory: %s", dirname);
+    LOG_TRACE("List directory: %s", dirname);
 
     paths = calloc(10, sizeof(*paths));
     check_oom(paths);
